pyproject.toml
[project]
name = "daily3albums"
version = "0.1.0"
description = "Triangulum: Daily 3 Albums (static site + scheduled generation)"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "pillow>=10.4.0",
  "rapidfuzz>=3.9.0",
  "boto3>=1.34.0",
]

[project.scripts]
daily3albums = "daily3albums.cli:main"

[tool.setuptools]
packages = ["daily3albums"]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"